% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-constructors.R
\name{bbm_sim_class}
\alias{bbm_sim_class}
\title{The 'bbm_sim' class.}
\usage{
bbm_sim_class(tbl, sim_D, sim_dim, bleach_rate = 0)
}
\arguments{
\item{tbl}{An appropriate \link[tibble:tibble]{tibble}.}

\item{sim_D}{The diffusivity constant used during the simulation which
created \code{tbl}.}

\item{sim_dim}{The dimensions of the space used during the simulation which
created \code{tbl}.}

\item{bleach_rate}{The bleaching rate (see \code{\link[=bbm_bleach_simulation]{bbm_bleach_simulation()}}).}
}
\value{
An object of class \code{bbm_sim}.
}
\description{
An object of class \code{bbm_sim} is a \link[tibble:tibble]{tibble} with \code{d + 2}
columns, where \code{d} is the number of dimensions in the space of the
simulation. One column for time, one for the particle ID number and the other
\code{d} for the positions of the particles in \code{d}-dimensional space.
\itemize{\item Column 1 has name \code{t} and is of type \code{<dbl>}. \item Column 2
has name \code{id} and is of type \code{<int>}. \item The rest of the columns are named
\code{x1}, \code{x2}, \code{x3} and so on and are of type \code{<dbl>}.} There are three necessary
attributes \code{sim_D}, \code{sim_dim} and \code{bleach_rate} which tell you the diffusivity constant,
the dimensions of the space used during the simulation and the bleaching rate (see \code{\link[=bbm_bleach_simulation]{bbm_bleach_simulation()}}).
}
\examples{
tbl <- tibble::tibble(t = 0, id = 1L, x1 = 0.5)
str(bbm_sim_class(tbl, sim_D = 1, sim_dim = 99.9))

}
