% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{bbm_simulate}
\alias{bbm_simulate}
\title{Simulate bounded brownian motion.}
\usage{
bbm_simulate(n_particles, D, dim, time_interval = 1, end_time,
  init_pos = NULL)
}
\arguments{
\item{n_particles}{A natural number. The number of particles to use in the
simulation.}

\item{D}{The diffusivity constant.}

\item{dim}{A vector giving the dimensions of the simulation space.}

\item{time_interval}{The time interval between observations.}

\item{end_time}{The time at which the simulation should end. The last
timpoint in the simulation will be before or at this time.}

\item{init_pos}{An \eqn{n} x d matrix giving the initial positions of the \code{n}
particles. If this is not set, then a random configuration is generated.}
}
\value{
An object of class \code{bbm_sim}: a \link[tibble:tibble]{tibble} with \code{d + 2} columns, where \code{d} is the number of dimensions in the space of the
simulation. One column for time, one for the particle ID number and the
other \code{d} for the positions of the particles in \code{d}-dimensional space.
}
\description{
Given the details of a rectangular boundary in 1, 2 or 3 dimensions, and a
diffusion constant, simulate the brownian trajectories of particles.
}
\examples{
bbm_simulate(n_particles = 2, D = 1.5, dim = c(3, 5), end_time = 2)

}
\seealso{
\code{\link[=bbm_simulate_img]{bbm_simulate_img()}}
}
