% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{bbm_simulation_to_img}
\alias{bbm_simulation_to_img}
\title{Create an image series from a \code{bbm_sim}.}
\usage{
bbm_simulation_to_img(bbms, pixel_size, method = c("count", "poisson"),
  brightness = NULL, check_duplicates = TRUE)
}
\arguments{
\item{bbms}{An object of class \link[=bbm_sim_class]{bbm_sim}; the result of a
simulation with \code{\link[=bbm_sim]{bbm_sim()}}. This must be 2-dimensional.}

\item{pixel_size}{The pixel size for the image. A grid of pixels is filled
out over the simulation space, starting at the bottom-left (the origin).
Any pixels crossing the boundary of the simulation are discarded.}

\item{method}{How shall the pixel values be calculated? \code{method = "count"}
results in each pixel value being a count of the number of particles
therein. \code{method = "poisson"} results in each particle being treated as a
Poisson-emmitter of photons with rate specified in the \code{brightness}
parameter.}

\item{brightness}{Required only if \code{method = poisson}. The poisson photon
emmission rate of the diffusing particles in units of \emph{per frame}. This is
the brightness 'epsilon' referred to in the field of number and brightness,
a method in fluorescence fluctuation spectroscopy.}

\item{check_duplicates}{Check for duplicate coordinates at each time point of the
simulation, throwing an error if any are present. One reason to turn this
off is for speed; the check slows things down slightly. If it's off and
\code{df} does contain duplicate coordinates, the value associated with the last
duplicate coordinate will count.}
}
\value{
An \link[ijtiff:ijtiff_img]{ijtiff_img}.
}
\description{
Given an output of \code{\link[=bbm_sim]{bbm_sim()}}, create an image in the style of an
\link[ijtiff:ijtiff_img]{ijtiff_img} where each frame represents a time-point in
the simulation. You must set the pixel size and there is the option for the
pixel values to be counts of the number of molecules in that pixel, or
Poisson photon counts where each particle has a user-defined brightness.
}
\examples{
sim <- bbm_simulate(n_particles = 2, D = 1.5, dim = c(3, 5), end_time = 2)
bbm_simulation_to_img(sim, pixel_size = 1, method = "c")
bbm_simulation_to_img(sim, pixel_size = 1, method = "p", brightness = 9)

}
\seealso{
\link{bbm_simulate_img}
}
