% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bleach_fraction_to_rate}
\alias{bleach_fraction_to_rate}
\title{Convert a bleaching fraction to a bleaching rate}
\usage{
bleach_fraction_to_rate(frac, n_frames)
}
\arguments{
\item{frac}{A number. The fraction of bleaching desired for the final
time-point relative to the first. So if you want 20% bleaching by the final
time-point (i.e. 80% of fluorescence remains), set \code{frac = 0.2}.}

\item{n_time_points}{A natural number. The number of time-points in the
simulation.}
}
\value{
A number.
}
\description{
Given a bleaching simulation of \code{n_time_points} time-points, this function
tells you the bleaching rate that will result in the final time-point having
bleached by a fraction \code{frac} relative to the first time-point
}
\examples{
bleach_fraction_to_rate(0.2, 1000)  # 20\% bleaching over 1000 time-points

}
