% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coords_to_arr}
\alias{coords_to_arr}
\title{Convert a data frame of coordinates and values to an array.}
\usage{
coords_to_arr(df, missing_value = NA, check_duplicates = TRUE)
}
\arguments{
\item{df}{A data frame whose columns are named \code{x1}, \code{x2}, \code{x3}, . . . , \code{xn}
and \code{value} (there can be no other columns). The \code{x} columns must be
positive integers.}

\item{missing_value}{Coordinates of \code{arr} not specified in \code{df} are set to
this value (default \code{NA}).}

\item{check_duplicates}{Check for duplicate coordinates in the data frame \code{df},
throwing an error if any are present. One reason to turn this off is for
speed; the check slows things down slightly. If it's off and \code{df} does
contain duplicate coordinates, the value associated with the last duplicate
coordinate will count.}
}
\value{
An \code{n} dimensional array with the same type as \code{df$value}.
}
\description{
Given a data frame with columns \code{x1}, \code{x2}, \code{x3}, . . . , \code{xn} and a column
\code{value}, create an array \code{arr} of dimension \code{n} where \code{arr[x1, x2, . . ., xn] = value}.
}
\examples{
coords <- tibble::tribble(~x1, ~x2, ~value,
                            1,   1,      3,
                            1,   2,      7,
                            2,   2,      0)
coords_to_arr(coords)
coords <- tibble::tribble(~x1, ~x2, ~ x3, ~value,
                            1,   1,    3,      3,
                            1,   2,    2,      7,
                            2,   2,    1,      0)
coords_to_arr(coords)
}
